pyproject.toml
[project]
name = "osiris-be"
version = "0.1.0"
description = "Proyecto Osiris Backend"
authors = [
    {name = "Open Latina", email = "openlatina.cue@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
packages = [{ include = "osiris", from = "src" }]
dependencies = [
    "fe-ec @ lib/fe_ec-0.1.0-py3-none-any-3.whl",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "fastapi (>=0.120.0,<0.130.0)",
    "starlette (>=0.49.1,<0.50.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlmodel (>=0.0.22)",
    "bcrypt (>=4.0.1,<5.0.0)",
    "psycopg (>=3.2.12,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "urllib3 (>=2.6.3,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
httpx = "^0.27"
ruff = "^0.5"            # lint
mypy = "^1.11"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
norecursedirs = ["tests/live_smoke"]
python_files = ["test_*.py"]
pythonpath = [".", "src"]
markers = [
    "smoke: smoke tests de flujos críticos E0-E5",
    "live_smoke: smoke tests contra servidor HTTP real",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
